+++ packages/dartcv/test/core/mat_test.dart
@@ -3,6 +3,7 @@
 import 'dart:ffi' as ffi;
 
 import 'package:dartcv4/dartcv.dart' as cv;
+import 'package:ffi/ffi.dart';
 import 'package:test/test.dart';
 
 void main() async {
@@ -92,6 +93,30 @@ void main() async {
     }
   });
 
+  test('Mat.fromBuffer', () {
+    const int rows = 3;
+    const int cols = 3;
+    const int channels = 3;
+    const int len = rows * cols * channels;
+    final pixels = List.generate(len, (i) => i);
+
+    for (var i = 0; i < 100; i++) {
+      final buff = calloc<ffi.Uint8>(len);
+      buff.asTypedList(len).setAll(0, pixels);
+
+      final mat = cv.Mat.fromBuffer(
+        rows,
+        cols,
+        const cv.MatType.CV_8UC(channels),
+        buff.cast(),
+      );
+      expect(mat.isEmpty, false);
+      expect(mat.shape, [rows, cols, channels]);
+      expect(mat.at<cv.Vec3b>(0, 0), cv.Vec3b(0, 1, 2));
+      calloc.free(buff);
+    }
+  });
+
   test('Mat.fromVec', () {
     const int rows = 3;
     const int cols = 3;
@@ -99,74 +124,76 @@ void main() async {
 
     final data = List.generate(rows * cols * channels, (i) => i);
 
-    {
-      const type = cv.MatType.CV_8UC(channels);
-      final mat = cv.Mat.fromVec(data.vecUChar, rows: rows, cols: cols, type: type);
-      expect(mat.isEmpty, false);
-      expect(mat.shape, [rows, cols, channels]);
-      expect(mat.at<cv.Vec3b>(0, 0), cv.Vec3b(0, 1, 2));
+    for (final copyData in [true, false]) {
+      {
+        const type = cv.MatType.CV_8UC(channels);
+        final mat = cv.Mat.fromVec(data.vecUChar, rows: rows, cols: cols, type: type, copyData: copyData);
+        expect(mat.isEmpty, false);
+        expect(mat.shape, [rows, cols, channels]);
+        expect(mat.at<cv.Vec3b>(0, 0), cv.Vec3b(0, 1, 2));
 
-      expect(mat.toFmtString(fmtType: cv.Mat.FMT_NUMPY), """
+        expect(mat.toFmtString(fmtType: cv.Mat.FMT_NUMPY), """
 array([[[  0,   1,   2], [  3,   4,   5], [  6,   7,   8]],
        [[  9,  10,  11], [ 12,  13,  14], [ 15,  16,  17]],
        [[ 18,  19,  20], [ 21,  22,  23], [ 24,  25,  26]]], dtype='uint8')""");
-    }
+      }
 
-    {
-      const type = cv.MatType.CV_8SC(channels);
-      final mat = cv.Mat.fromVec(data.vecChar, rows: rows, cols: cols, type: type);
-      expect(mat.isEmpty, false);
-      expect(mat.shape, [rows, cols, channels]);
-      expect(mat.at<cv.Vec3b>(0, 1), cv.Vec3b(3, 4, 5));
-    }
+      {
+        const type = cv.MatType.CV_8SC(channels);
+        final mat = cv.Mat.fromVec(data.vecChar, rows: rows, cols: cols, type: type, copyData: copyData);
+        expect(mat.isEmpty, false);
+        expect(mat.shape, [rows, cols, channels]);
+        expect(mat.at<cv.Vec3b>(0, 1), cv.Vec3b(3, 4, 5));
+      }
 
-    {
-      const type = cv.MatType.CV_16UC(channels);
-      final mat = cv.Mat.fromVec(data.u16, rows: rows, cols: cols, type: type);
-      expect(mat.isEmpty, false);
-      expect(mat.shape, [rows, cols, channels]);
-      expect(mat.at<cv.Vec3w>(0, 1), cv.Vec3w(3, 4, 5));
-    }
+      {
+        const type = cv.MatType.CV_16UC(channels);
+        final mat = cv.Mat.fromVec(data.u16, rows: rows, cols: cols, type: type, copyData: copyData);
+        expect(mat.isEmpty, false);
+        expect(mat.shape, [rows, cols, channels]);
+        expect(mat.at<cv.Vec3w>(0, 1), cv.Vec3w(3, 4, 5));
+      }
 
-    {
-      const type = cv.MatType.CV_16SC(channels);
-      final mat = cv.Mat.fromVec(data.i16, rows: rows, cols: cols, type: type);
-      expect(mat.isEmpty, false);
-      expect(mat.shape, [rows, cols, channels]);
-      expect(mat.at<cv.Vec3s>(0, 1), cv.Vec3s(3, 4, 5));
-    }
+      {
+        const type = cv.MatType.CV_16SC(channels);
+        final mat = cv.Mat.fromVec(data.i16, rows: rows, cols: cols, type: type, copyData: copyData);
+        expect(mat.isEmpty, false);
+        expect(mat.shape, [rows, cols, channels]);
+        expect(mat.at<cv.Vec3s>(0, 1), cv.Vec3s(3, 4, 5));
+      }
 
-    {
-      const type = cv.MatType.CV_32SC(channels);
-      final mat = cv.Mat.fromVec(data.i32, rows: rows, cols: cols, type: type);
-      expect(mat.isEmpty, false);
-      expect(mat.shape, [rows, cols, channels]);
-      expect(mat.at<cv.Vec3i>(0, 1), cv.Vec3i(3, 4, 5));
-    }
+      {
+        const type = cv.MatType.CV_32SC(channels);
+        final mat = cv.Mat.fromVec(data.i32, rows: rows, cols: cols, type: type, copyData: copyData);
+        expect(mat.isEmpty, false);
+        expect(mat.shape, [rows, cols, channels]);
+        expect(mat.at<cv.Vec3i>(0, 1), cv.Vec3i(3, 4, 5));
+      }
 
-    {
-      const type = cv.MatType.CV_32FC(channels);
-      final mat = cv.Mat.fromVec(data.f32, rows: rows, cols: cols, type: type);
-      expect(mat.isEmpty, false);
-      expect(mat.shape, [rows, cols, channels]);
-      expect(mat.at<cv.Vec3f>(0, 1), cv.Vec3f(3, 4, 5));
-    }
+      {
+        const type = cv.MatType.CV_32FC(channels);
+        final mat = cv.Mat.fromVec(data.f32, rows: rows, cols: cols, type: type, copyData: copyData);
+        expect(mat.isEmpty, false);
+        expect(mat.shape, [rows, cols, channels]);
+        expect(mat.at<cv.Vec3f>(0, 1), cv.Vec3f(3, 4, 5));
+      }
 
-    {
-      const type = cv.MatType.CV_64FC(channels);
-      final mat = cv.Mat.fromVec(data.f64, rows: rows, cols: cols, type: type);
-      expect(mat.isEmpty, false);
-      expect(mat.shape, [rows, cols, channels]);
-      expect(mat.at<cv.Vec3d>(0, 1), cv.Vec3d(3, 4, 5));
-    }
+      {
+        const type = cv.MatType.CV_64FC(channels);
+        final mat = cv.Mat.fromVec(data.f64, rows: rows, cols: cols, type: type, copyData: copyData);
+        expect(mat.isEmpty, false);
+        expect(mat.shape, [rows, cols, channels]);
+        expect(mat.at<cv.Vec3d>(0, 1), cv.Vec3d(3, 4, 5));
+      }
 
-    {
-      const type = cv.MatType.CV_16FC(channels);
-      final mat = cv.Mat.fromVec(data.f16, rows: rows, cols: cols, type: type);
-      expect(mat.isEmpty, false);
-      expect(mat.shape, [rows, cols, channels]);
-      final pix = mat.at<cv.Vec3w>(0, 1);
-      expect(pix.val, [3.0.fp16, 4.0.fp16, 5.0.fp16]);
+      {
+        const type = cv.MatType.CV_16FC(channels);
+        final mat = cv.Mat.fromVec(data.f16, rows: rows, cols: cols, type: type, copyData: copyData);
+        expect(mat.isEmpty, false);
+        expect(mat.shape, [rows, cols, channels]);
+        final pix = mat.at<cv.Vec3w>(0, 1);
+        expect(pix.val, [3.0.fp16, 4.0.fp16, 5.0.fp16]);
+      }
     }
   });